chore: rename component's file
